+++ Código/Classes/Business/Produtos/OpcoesAdicionais.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Produtos;
 
 public enum OpcoesAdicionais {
     PEPPERONI(4),
