+++ Código/Classes/Business/Produtos/OpcoesComida.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Produtos;
 
 public enum OpcoesComida {
     PRATO_FEITO(15),
