+++ Código/Classes/Business/Produtos/ProdutosComAdicionais.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Produtos;
 
 import java.util.HashSet;
 import java.util.Set;
