+++ Código/Classes/Business/Venda/Cliente.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Venda;
 
 import java.io.Serializable;
 import java.time.LocalDate;
