+++ Código/Classes/Business/Venda/FBranco.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Venda;
 
 public class FBranco implements IFidelidade {
     @Override
