+++ Código/Classes/Business/Venda/IFidelidade.java
@@ -1,4 +1,4 @@
-package Business;
+package Business.Venda;
 
 import java.io.Serializable;
